style(history): change weekday format from long to short
The weekday format in the date display was updated from 'long' to 'short' to show abbreviated day names (e.g., Mon, Tue) for better readability and consistency with common date formatting practices.

+++ app/(tabs)/history.tsx
@@ -152,9 +152,9 @@ export default function HistoryScreen() {
       return 'Yesterday';
     }
     return date.toLocaleDateString('en-US', { 
-      weekday: 'long',
+      weekday: 'short', // Changed from 'long' to 'short' to show abbreviated day names (Mon, Tue, etc.)
       year: 'numeric',
-      month: 'long',
+      month: 'short',
       day: 'numeric'
     });
   };
